+++ test/models/housekeeping_test.rb
@@ -4,7 +4,10 @@ class HousekeepingTest < ActiveSupport::TestCase
   setup do
     # Create victim tables and triggers
     create_victim_structures
-    run_with_current_user { create_event_logs_for_test(11) }                    # ensure that at least one interval partition is created
+    run_with_current_user do
+      EventLog.adjust_interval                                                  # Restore possibly wrong setting on table
+      create_event_logs_for_test(11)                                            # ensure that at least one interval partition is created
+    end
   end
 
   # Use the DB time instead of Time.now to avoid timezone conflicts
@@ -17,7 +20,7 @@ class HousekeepingTest < ActiveSupport::TestCase
   def assure_last_partition
     case MovexCdc::Application.config.db_type
     when 'ORACLE' then
-      if MovexCdc::Application.partitioning?
+      if MovexCdc::Application.partitioning? && DatabaseOracle.db_version < '12.2'
         part_count = Database.select_one("SELECT COUNT(*) FROM User_Tab_Partitions WHERE Table_Name = 'EVENT_LOGS'")
         if part_count < 2
           Database.select_all("SELECT Partition_Name, High_Value, Interval FROM User_Tab_Partitions WHERE Table_Name = 'EVENT_LOGS' ORDER BY Partition_Position").each do |part|
@@ -84,6 +87,7 @@ class HousekeepingTest < ActiveSupport::TestCase
       Database.execute "ALTER TABLE Event_Logs SET INTERVAL (NUMTODSINTERVAL(60,'SECOND'))" if Database.db_version < '13' # Workaround bug in 12.1.0.2 where oldest range partition cannot be dropped if split is done with older high_value (younger partition can be dropped instead)
       log_partition_state(false, 'set_high_value_time after splitting')
     end
+    force_interval_partition_creation(high_value_time+interval)                 # create a second partition directly after the first partition (will be with interval = NO)
     EventLog.adjust_interval                                                    # adjust in DB according to MovexCdc::Application.config.partition_interval
     force_interval_partition_creation(last_partition_time)                      # ensure existence of at least one interval partition
   end
@@ -170,6 +174,7 @@ class HousekeepingTest < ActiveSupport::TestCase
       if MovexCdc::Application.partitioning?
         # Drop all partitions except the first one (possibly there are only two range partitions at this point
         initial_partiton_count = drop_all_event_logs_partitions_except_1
+        Rails.logger.debug('HousekeepingTest.do_housekeeping with locked partition') { "initial_partiton_count = #{initial_partiton_count}" }
         last_part = Database.select_first_row("SELECT Partition_Name, High_Value
                              FROM   User_Tab_Partitions
                              WHERE  Table_Name = 'EVENT_LOGS'
@@ -194,8 +199,9 @@ class HousekeepingTest < ActiveSupport::TestCase
 
           raise("Housekeeping not finished until limit") if hk_thread.join(30).nil?
           end_partition_count = Database.select_one("SELECT COUNT(*) FROM User_Tab_Partitions WHERE Table_Name = 'EVENT_LOGS'")
-          # There should remain: the first partition, three partitions with pending inserts and the last partition
-          assert_equal 4 + initial_partiton_count, end_partition_count, log_on_failure("Temporary partition with pending insert should not be deleted. Current interval = #{MovexCdc::Application.config.partition_interval}")
+          # There should remain: the first partition (if < 12.2), three partitions with pending inserts and the last partition
+          remaining_partitions = (DatabaseOracle.db_version < '12.2' ? 5 : 4)
+          assert_equal remaining_partitions, end_partition_count, log_on_failure("Temporary partition with pending insert should not be deleted. Current interval = #{MovexCdc::Application.config.partition_interval}")
         end
         Database.execute "DELETE FROM Event_Logs"                               # Ensure all unprocessable records are removed
         restore_partitioning
@@ -213,12 +219,19 @@ class HousekeepingTest < ActiveSupport::TestCase
             drop_all_event_logs_partitions_except_1                             # possible need to remove the first partition if only range partitions exist
             max_seconds_for_interval_prev= 700000*prev_interval                 # > 1/2 of max. partition count (1024*1024-1) for default interval
             set_high_value_time(time_now-max_seconds_for_interval_prev, prev_interval, time_now) # set old high_value to 1/2 of possible partition count and default interval
-            Housekeeping.get_instance.check_partition_interval
-            log_partition_state(false, 'After check_partitions')                                               # log partitions
+            Housekeeping.get_instance.do_housekeeping                           # Regular drop of partitions, ensures drop for >= 12.2
+            Housekeeping.get_instance.check_partition_interval                  # Test for < 12.2
+            max_expected_seconds_for_interval = (1024*1024)*10*interval         # < 1/4 of max. partition count (1024*1024-1) for interval
+            min_expected_hv = time_now-max_expected_seconds_for_interval
 
             current_hv = get_time_from_high_value(1)
-            max_expected_seconds_for_interval = (1024*1024)*10*interval          # < 1/4 of max. partition count (1024*1024-1) for interval
-            min_expected_hv = time_now-max_expected_seconds_for_interval
+            if current_hv <= min_expected_hv                                    # Repeat check_partition_interval to lift the second old partition also
+              Housekeeping.get_instance.check_partition_interval
+              current_hv = get_time_from_high_value(1)
+            end
+
+            log_partition_state(false, 'After check_partitions')                                               # log partitions
+
             assert current_hv > min_expected_hv, log_on_failure("high value now (#{current_hv}) should be younger than 1/4 related to max. partition count (1024*1024-1) for interval #{interval} seconds (#{min_expected_hv})")
           end
 
@@ -238,6 +251,11 @@ class HousekeepingTest < ActiveSupport::TestCase
         restore_partitioning
       end
     end
+  end
+
+  test "check_partition_interval with locked middle partition" do
+    # TODO: Test that housekeeping fails if the two first partitions are without a gap and the second partition is locked by another transaction
+    # The first partition should be old enough to be caught by check_partition_interval
   end
 
   test "check_partition_interval with no gaps between min and last partition" do